Generate randomly some text noise
I'm not sure if this even needed at the end. For now it makes possible to generate text with trained models without needing actual training data or primer text, just trained model and you're good to go. Example of generated text from random strings:

I of the sin.
And dead and passion.
And they think to be free.
I tried to be not to be weaker.
The sorrow shall dwell the dark they will hold to the fate.
And the souls fly through brains.
Heartless thirsting meaning.
With any scar and under the burning stars.
You sold to observe.
I will run in started force.
A compassion of darkness is for an and souls let the sign.
I am the path i am shame to.